# Conflicts:
#	app/client/src/entities/FeatureFlags.ts

+++ app/client/cypress/support/e2e.js
@@ -43,6 +43,13 @@ Cypress.on("uncaught:exception", () => {
 });
 
 Cypress.on("fail", (error) => {
+  cy.window()
+    .its("store")
+    .invoke("getState")
+    .then((state) => {
+      cy.log(`Editor initialised: ${state.ui.editor.initialized}`);
+      cy.log(`Loading guided tour: ${state.ui.guidedTour.loading}`);
+    });
   throw error; // throw error to have test still fail
 });
 
@@ -78,7 +85,7 @@ before(function () {
   cy.window().then((window) => {
     window.indexedDB.deleteDatabase("Appsmith");
   });
-  cy.visit("/setup/welcome");
+  cy.visit("/setup/welcome", { timeout: 60000 });
   cy.wait("@getMe");
   cy.wait(2000);
   cy.url().then((url) => {
@@ -169,11 +176,14 @@ after(function () {
   if (RapidMode.config.enabled) {
     return;
   }
-  //-- Deleting the application by Api---//
-  cy.DeleteAppByApi();
-  //-- LogOut Application---//
-  cy.LogOut();
-
+  if (Cypress.env("CYPRESS_CI") === "1") {
+    //-- Deleting the application by Api---//
+    cy.DeleteAppByApi();
+    //-- LogOut Application---//
+    cy.LogOut();
+  } else {
+    cy.log("Running locally, hence skipping app delete!");
+  }
   // Commenting until Upgrade Appsmith cases are fixed
   // const tedUrl = "http://localhost:5001/v1/parent/cmd";
   // cy.log("Start the appsmith container");
